more msvc warnings

+++ ultimaille/helpers/disjointset.h
@@ -52,7 +52,7 @@ namespace UM {
         }
 
         int size() const {
-            return m_ids.size();
+            return static_cast<int>(m_ids.size());
         }
 
         // return the number of sets
